Separate `containers` from `components`
